Add channel messaging demo with two iframes
This commit adds a new HTML file, index.html, which contains a channel messaging demo with two iframes. The iframes load page1.html and page2.html respectively. The demo demonstrates how to use MessageChannel to communicate between the main frame and the iframes. Messages are sent from the main frame to the iframes, and the iframes respond by sending messages back to the main frame. The received messages are displayed in the output elements on the main frame.